Data_Store_v2

+++ src/components/data_storing_api.py
@@ -5,7 +5,7 @@ from src.logger import logging
 from src.constants.file_constants import INTERACTIONS_PARQUET_FILEPATH , COURSES_PARQUET_FILEPATH, ENTITY_COURSES_PARQUET_FILEPATH, ENTITY_INTERACTIONS_PARQUET_FILEPATH
 from csv import writer
 from src.configurations.mongo_config import MongoDBClient
-
+from src.components.data_validation import DataValidation
 class StoreData:
     '''
     We take courses data, users data, user-course interactions data from apis, put it in proper and required format
@@ -16,9 +16,9 @@ class StoreData:
             self.mongo_client = MongoDBClient()
             self.connection = self.mongo_client.dbcollection
             self.index_connection = self.mongo_client.index_collection
-            self.reverse_course_connection = self.mongo_client.reverse_courses_collection_name
-
-            #self.data_validation = DataValidation()
+            self.course_connection = self.mongo_client.course_collection
+            #self.reverse_course_connection = self.mongo_client.reverse_courses_collection_name
+            self.data_validation = DataValidation()
             pass
         except Exception as e:
             raise DataException(e,sys)
@@ -33,20 +33,20 @@ class StoreData:
             #Create current timestamp for the interaction
             logging.info("Getting current UTC Timestamp")
             current_timestamp = pd.Timestamp.now()
-            current_timestamp = current_timestamp.strftime("%Y-%m-%d %H:%M:%S").to_list()
+            current_timestamp = current_timestamp.strftime("%Y-%m-%d %H:%M:%S")
 
         
-            #data_validation_status = False
+            data_validation_status = False
 
-            #data_validation_status = self.data_validation.check_courses_validation(item_dict) #return true if everything is valid
+            data_validation_status = self.data_validation.check_courses_validation(item_dict) #return true if everything is valid
 
-            #if data_validation_status == False:
-            #    return False
+            if data_validation_status == False:
+                return False
                             
             data_validation_status = True
 
             logging.info("Getting latest course feature id from mongo db")
-            latest_course_id = self.index_connection.find({"id_name":"latest_course_feature_id"},{'_id': 0, 'value':1}).next().get('value')
+            latest_course_id = dict(self.index_connection.find({"id_name":"latest_course_feature_id"},{'_id': 0, 'value':1}).next()).get('value')
             new_course_feature_id = latest_course_id + 1
             data_index = { "id_name":"latest_course_feature_id", "value": new_course_feature_id}
             self.index_connection.delete_one({"id_name":"latest_course_feature_id"})
@@ -75,22 +75,23 @@ class StoreData:
                 tags += "Cloud "
                 self.connection.insert_one({"category": "cloud", "course_name": item_dict["course_name"]})
 
-            self.mongo_client.close()
-
             new_df = { 
-                "event_timestamp" : current_timestamp,
+                "event_timestamp" : [current_timestamp],
                 "course_feature_id" : [new_course_feature_id],
-                "course_id" : [new_course_feature_id],
+                "course_id" : [new_course_feature_id - 1],
                 "course_name" : [str(item_dict["course_name"])],
                 "course_tags" : [tags],
             }
             new_course_df = pd.DataFrame(new_df)
+            new_course_df["event_timestamp"] = pd.to_datetime(new_course_df["event_timestamp"])
+
 
             new_df_entity = { 
                 "event_timestamp" : current_timestamp,
                 "course_feature_id" : [new_course_feature_id],
              }
             new_course_df_entity = pd.DataFrame(new_df_entity)
+            new_course_df_entity["event_timestamp"] = pd.to_datetime(new_course_df_entity["event_timestamp"])
 
             logging.info("Storing new feature data in parquet")
             dfd = pd.read_parquet(COURSES_PARQUET_FILEPATH)
@@ -101,7 +102,7 @@ class StoreData:
             logging.info("Storing new entity data in parquet")
             dfd = pd.read_parquet(ENTITY_COURSES_PARQUET_FILEPATH)
             dfd = dfd.append(new_course_df_entity, ignore_index=True)
-            dfd.to_parquet(ENTITY_COURSES_PARQUET_FILEPATH, index=False)
+            dfd.to_parquet(ENTITY_COURSES_PARQUET_FILEPATH, index=False)        
 
             logging.info("Exiting the store_courses_data function of StoreData class")
 
@@ -120,14 +121,14 @@ class StoreData:
             #Create current timestamp for the interaction
             logging.info("Getting current UTC Timestamp")
             current_timestamp = pd.Timestamp.now()
-            current_timestamp = current_timestamp.strftime("%Y-%m-%d %H:%M:%S").to_list()
+            current_timestamp = current_timestamp.strftime("%Y-%m-%d %H:%M:%S")
 
-            #data_validation_status = False
+            data_validation_status = False
 
-            #data_validation_status = self.data_validation.check_interactions_validation(item_dict) #return true if everything is valid
+            data_validation_status = self.data_validation.check_interactions_validation(item_dict) #return true if everything is valid
 
-            #if data_validation_status == False:
-            #    return False
+            if data_validation_status == False:
+                return False
             
             data_validation_status = True
 
@@ -144,11 +145,11 @@ class StoreData:
 
             logging.info("Finding the course id")
             #Find Course ID of the course taken
-            find_course_id = self.reverse_course_connection.find({"course_name" : course_taken},{'_id':0,'course_id':1}).next().get('course_id')
+            find_course_id = dict(self.course_connection.find({"course_name":course_taken},{'_id':0, 'course_id':1}).next()).get('course_id')
 
             if event_viewed:
                 logging.info("Getting latest interaction id from mongo db")
-                latest_interaction_id = self.index_connection.find({"id_name":"latest_interaction_id"},{'_id': 0, 'value':1}).next().get('value')
+                latest_interaction_id = dict(self.index_connection.find({"id_name":"latest_interaction_id"},{'_id': 0, 'value':1}).next()).get('value')
                 new_interaction_id = latest_interaction_id + 1
                 data_index = { "id_name":"latest_interaction_id", "value": new_interaction_id}
                 self.index_connection.delete_one({"id_name":"latest_interaction_id"})
@@ -158,22 +159,24 @@ class StoreData:
                 "user_id" : [user_id],
                 "event" : 1,
                 "course_id" : [find_course_id],
-                "event_timestamp" : current_timestamp,
+                "event_timestamp" : [current_timestamp],
                 "interaction_id" : [new_interaction_id],
                 }
                 view_df = pd.DataFrame(new_df_viewed)
+                view_df["event_timestamp"] = pd.to_datetime(view_df["event_timestamp"])
                 complete_df = pd.concat([complete_df,view_df], axis=0)
 
                 entity_df_viewed = { 
-                "event_timestamp" : current_timestamp,
+                "event_timestamp" : [current_timestamp],
                 "interaction_id" : [new_interaction_id],
                 }
                 view_df_entity = pd.DataFrame(entity_df_viewed)
+                view_df_entity["event_timestamp"] = pd.to_datetime(view_df_entity["event_timestamp"])
                 complete_df_entity = pd.concat([complete_df_entity,view_df_entity], axis=0)
 
             if event_wishlisted:
                 logging.info("Getting latest interaction id from mongo db")
-                latest_interaction_id = self.index_connection.find({"id_name":"latest_interaction_id"},{'_id': 0, 'value':1}).next().get('value')
+                latest_interaction_id = dict(self.index_connection.find({"id_name":"latest_interaction_id"},{'_id': 0, 'value':1}).next()).get('value')
                 new_interaction_id = latest_interaction_id + 1
                 data_index = { "id_name":"latest_interaction_id", "value": new_interaction_id}
                 self.index_connection.delete_one({"id_name":"latest_interaction_id"})
@@ -183,22 +186,24 @@ class StoreData:
                 "user_id" : [user_id],
                 "event" : 2,
                 "course_id" : [find_course_id],
-                "event_timestamp" : current_timestamp,
+                "event_timestamp" : [current_timestamp],
                 "interaction_id" : [new_interaction_id],
                 }
                 wishlisted_df = pd.DataFrame(new_df_wishlisted)
+                wishlisted_df["event_timestamp"] = pd.to_datetime(wishlisted_df["event_timestamp"])
                 complete_df = pd.concat([complete_df,wishlisted_df], axis=0)
 
                 entity_df_wishlisted = { 
-                "event_timestamp" : current_timestamp,
+                "event_timestamp" : [current_timestamp],
                 "interaction_id" : [new_interaction_id],
                 }
                 wish_df_entity = pd.DataFrame(entity_df_wishlisted)
+                wish_df_entity["event_timestamp"] = pd.to_datetime(wish_df_entity["event_timestamp"])
                 complete_df_entity = pd.concat([complete_df_entity,wish_df_entity], axis=0)
                 
             if event_enrolled:
                 logging.info("Getting latest interaction id from mongo db")
-                latest_interaction_id = self.index_connection.find({"id_name":"latest_interaction_id"},{'_id': 0, 'value':1}).next().get('value')
+                latest_interaction_id = dict(self.index_connection.find({"id_name":"latest_interaction_id"},{'_id': 0, 'value':1}).next()).get('value')
                 new_interaction_id = latest_interaction_id + 1
                 data_index = { "id_name":"latest_interaction_id", "value": new_interaction_id}
                 self.index_connection.delete_one({"id_name":"latest_interaction_id"})
@@ -208,17 +213,19 @@ class StoreData:
                 "user_id" : [user_id],
                 "event" : 3,
                 "course_id" : [find_course_id],
-                "event_timestamp" : current_timestamp,
+                "event_timestamp" : [current_timestamp],
                 "interaction_id" : [new_interaction_id],
                 }
                 enrolled_df = pd.DataFrame(new_df_enrolled)
+                enrolled_df["event_timestamp"] = pd.to_datetime(enrolled_df["event_timestamp"])
                 complete_df = pd.concat([complete_df,enrolled_df], axis=0)
 
                 entity_df_enrolled = { 
-                "event_timestamp" : current_timestamp,
+                "event_timestamp" : [current_timestamp],
                 "interaction_id" : [new_interaction_id],
                 }
                 enrolled_df_entity = pd.DataFrame(entity_df_enrolled)
+                enrolled_df_entity["event_timestamp"] = pd.to_datetime(enrolled_df_entity["event_timestamp"])
                 complete_df_entity = pd.concat([complete_df_entity,enrolled_df_entity], axis=0)
 
 
@@ -238,7 +245,7 @@ class StoreData:
 
         except Exception as e:
             raise DataException(e,sys)
-
+        
 
 if __name__ == "__main__":
     getdataobj  =  StoreData()
